feat: functional scripting with hot reloading

+++ Source/Modules/ModuleEditor.cpp
@@ -309,7 +309,7 @@ void ModuleEditor::RefreshInspector() const
 
 }
 
-std::pair<int, int> ModuleEditor::GetAvailableRegion()
+std::pair<float, float> ModuleEditor::GetAvailableRegion()
 {
 #ifdef ENGINE
 	ImVec2 region = scene->GetAvailableRegion();
